significant refactoring with AI
