+++ .github/workflows/build.yml
@@ -1,0 +1,59 @@
+name: Go Build and Deploy
+
+on:
+  push:
+    branches: [ "main" ]
+  pull_request:
+    branches: [ "main" ]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up Go
+        uses: actions/setup-go@v5
+        with:
+          go-version: '1.22'
+
+      - name: Build Binary
+        env:
+          GOARCH: amd64
+          GOOS: linux
+          CGO_ENABLED: 0
+        run: |
+          go mod download
+          go build -o bootstrap ./cmd/photo-lambda
+          zip function.zip bootstrap
+
+      - name: Upload Artifact
+        uses: actions/upload-artifact@v4
+        with:
+          name: lambda-package
+          path: function.zip
+
+# Ask JetBrains AI how to use OIDC instead of credentials like this
+#  deploy:
+#    if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+#    needs: build
+#    runs-on: ubuntu-latest
+#    steps:
+#      - name: Download Artifact
+#        uses: actions/download-artifact@v4
+#        with:
+#          name: lambda-package
+#
+#      - name: Configure AWS Credentials
+#        uses: aws-actions/configure-aws-credentials@v4
+#        with:
+#          aws-access-key-id: ${{ secrets.AWS_ACCESS_KEY_ID }}
+#          aws-secret-access-key: ${{ secrets.AWS_SECRET_ACCESS_KEY }}
+#          aws-region: us-east-1 # Change to your region
+#
+#      - name: Deploy to Lambda
+#        run: |
+#          aws lambda update-function-code \
+#            --function-name YourLambdaFunctionName \
+#            --zip-file fileb://function.zip
