+++ .gitignore
@@ -6,6 +6,3 @@
 **/*key
 
 **/*.zip
-
-photo-lambda
-bootstrap
