--- buildspec.yml
@@ -1,24 +1,0 @@
-version: 0.2
-
-phases:
-  install:
-    runtime-versions:
-      golang: 1.21
-  pre_build:
-    commands:
-      - echo Build started
-      - go mod download github.com/aws/aws-lambda-go
-      - go get github.com/TheBellman/photo-lambda
-      - GOARCH=amd64 GOOS=linux go test
-  build:
-    commands:
-      - go mod download github.com/aws/aws-lambda-go
-      - go get github.com/TheBellman/photo-lambda
-      - GOARCH=amd64 GOOS=linux go build -o bootstrap main.go image.go
-  post_build:
-    commands:
-      - echo Build completed
-artifacts:
-  files:
-    - bootstrap
-  name: photo-lambda-$(date +%Y%m%d.%H%M%s).zip
